fixed little bug & added accept/decline internship

+++ app/Http/Controllers/InternshipController.php
@@ -68,7 +68,7 @@ class InternshipController extends Controller
 
         $user = session()->get('user');
         if ($user->id != $internship->company_id) {
-            return redirect("/internships/$internship");
+            return redirect("/companies/".$internship->company_id);
         }
 
         $validation = Validator::make($request->all(), [
@@ -93,9 +93,15 @@ class InternshipController extends Controller
 
     public function delete($id){
         $internship = \App\Internship::where('id', $id)->first();
+
+        $user = session()->get('user');
+        if ($user->id != $internship->company_id) {
+            return redirect("/companies/".$internship->company_id."?internship=list");
+        }
+
         $company_id = $internship->company_id;
         $internship->delete();
-        
+            
         return redirect("/companies/$company_id?internship=list")->with('success', 'Internship successfully deleted!');
     }
 
@@ -154,5 +160,29 @@ class InternshipController extends Controller
         $internship->delete();
 
         return redirect("/internships/$id")->with('success', 'Successfully removed apply for internship!');
+    }
+
+    public function status($id) {
+        $student = $_GET["student"];
+        $status = $_GET["status"];
+        if (!empty($student) && !empty($status)) {
+            $request = \App\StudentInternship::where([
+                ['internship_id', $id],
+                ['student_id', $student]
+            ])->first();
+
+            if ($status == "accept") {
+                $state = 1;
+                $status = $status."ed";
+            } else {
+                $state = 2;
+                $status = $status."d";
+            }
+
+            $request->status = $state;
+            $request->save();
+            return redirect("/")->with('success', 'You successfully '.$status.' te request!');
+        }
+        return redirect("/")->with('error', 'Something went wrong!');
     }
 }
